// src/models/location/address.interface.ts
export interface Address{
	administrative_area_level_1_lng?: string;
	administrative_area_level_1_short?: string;
	administrative_area_level_2_lng?: string;
	administrative_area_level_2_short?: string;
	country_long: string;
	country_short: string;
	description: string;
	lat: number;
	lng: number;
	locality_lng?: string;
	locality_short?: string;
	name?: string;
	sublocality_lng?: string;
	sublocality_short?: string;
	vicinity?: string;
}